requirements.txt
Django==4.2.6
channels-redis==4.1.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
django-storages==1.14.1
django-extensions==3.2.3
django-simple-history==3.4.0
factory-boy==3.3.0
ipython==8.16.1
paramiko==3.3.1
pillow==10.0.1
pydifact==0.1.6
python-decouple==3.8
pytest-cov==4.1.0
pytest-django==4.5.2
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==3.0.0
pytest-asyncio==0.23.4
pandas==2.2.0
